Add example of using a raw DBObject style query
I spent a while working out how to do this, I'd totally missed the DBObject overload of the createQuery method and couldn't find an example of how to mix it with Morphia's built-in querying style. In fact, I'd started writing a "raw" addition to the "filter" method before I realised it must already be there.