Adjust projection generation when unwrapping single-argument expressions
This fixes the case where a single argument expression such as {$size: “$source”} generates an arg array
containing a string rather than a DBObject. The current behaviour would try to unwrap the “$source” and cast it to
a DBObject, which caused a ClassCastException.

By moving the check for single arguments to the toDBObject method, it’s possible to preserve the existing behaviour
where DBObjects are unwrapped and used as the argument while also correctly handling string arguments.